More utils 'improvements'

+++ visage_graphics/font.h
@@ -184,7 +184,7 @@ namespace visage {
     static PackedFont* loadPackedFont(int size, const std::string& file_path) {
       std::string id = "file: " + file_path + " - " + std::to_string(size);
       File file(file_path);
-      int file_size = 0;
+      size_t file_size = 0;
       std::unique_ptr<unsigned char[]> data = loadFileData(file, file_size);
       return instance()->createOrLoadPackedFont(id, size, data.get(), file_size);
     }
